Analog resolution modified to 10bit

+++ uCNC/src/hal/mcus/esp8266/mcu_esp8266.c
@@ -249,7 +249,7 @@ void mcu_disable_probe_isr(void)
  * can be defined either as a function or a macro call
  * */
 #ifndef mcu_get_analog
-uint8_t mcu_get_analog(uint8_t channel)
+uint16_t mcu_get_analog(uint8_t channel)
 {
 	return 0;
 }
